repozytorium pojazdóe cd

+++ src/model/MotorVehicle.cpp
@@ -29,6 +29,6 @@ double MotorVehicle::getActuallRentalPrice() {
     return actuallRentalPrice;
 }
 
-MotorVehicle::MotorVehicle(int baseRentPrice, std::string id, double engineDisplacement, double actuallRentalPrice)
+MotorVehicle::MotorVehicle(int baseRentPrice, std::string id, int engineDisplacement, double actuallRentalPrice)
         : Vehicle(baseRentPrice, id), engineDisplacement{engineDisplacement}, actuallRentalPrice{calculateActualRentalPrice()} {}
